+++ lib/src/tridy.dart
@@ -47,3 +47,40 @@ class Jidelnicek {
   List<Jidlo> jidla;
   Jidelnicek(this.den, this.jidla);
 }
+
+/// Reprezentuje informace o přihlášeném uživateli
+class Uzivatel {
+  /// Uživatelské jméno
+  String? uzivatelskeJmeno;
+
+  /// Jméno, jak je uvedené v základních údajích o uživateli
+  String? jmeno;
+
+  /// příjmení, jak je uvedené v základních údajích o uživateli
+  String? prijmeni;
+
+  /// Kategorie uživatele
+  String? kategorie;
+
+  /// Účet jídelny pro zasílání plateb
+  String? ucetProPlatby;
+
+  /// Variabilní symbol
+  String? varSymbol;
+
+  /// Specifický symbol
+  String? specSymbol;
+
+  /// Aktuální stav kreditu
+  double kredit;
+
+  Uzivatel(
+      {this.uzivatelskeJmeno,
+      this.jmeno,
+      this.prijmeni,
+      this.kategorie,
+      this.ucetProPlatby,
+      this.varSymbol,
+      this.kredit = 0.0,
+      this.specSymbol});
+}