+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0-alpha.2
+- Nevytvářet debugovací soubor
+- Místo ziskejKredit používáme ziskejUzivatele (Třída Uzivatel)
+- Requesty by měly vyhazovat Exception při chybném požadavku (status kódu)
 ## 0.1.0-alpha.1
 - Aktualizace licence